add useCases directories and add likeUsecases files for atomic db operations
