+++ src/usecases/likeUsecases/deleteLikeFromPlace.js
@@ -1,0 +1,3 @@
+const deleteLikeFromPlace = async (id) => await Place.findByIdAndDelete(id)
+
+export { deleteLikeFromPlace }
