+++ src/usecases/likeUsecases/deleteLikeFromRoute.js
@@ -1,0 +1,3 @@
+const deleteLikeFromRoute = async (id) => await Route.findByIdAndDelete(id)
+
+export { deleteLikeFromRoute }
