+++ src/usecases/likeUsecases/getLikesFromPlace.js
@@ -1,0 +1,3 @@
+const getLikesFromPlace = async (id) => await Place.findById(id)
+
+export { getLikesFromPlace }
