+++ src/usecases/likeUsecases/getLikesFromRoute.js
@@ -1,0 +1,3 @@
+const getLikesFromRoute = async (id) => await Route.findById(id)
+
+export { getLikesFromRoute }
