+++ src/usecases/likeUsecases/postLikeToPlace.js
@@ -1,0 +1,4 @@
+const postLikeToPlace = async (id, like) =>
+  await Place.findByIdAndUpdate(id, like)
+
+export { postLikeToPlace }
