+++ src/usecases/likeUsecases/postLikeToRoute.js
@@ -1,0 +1,4 @@
+const postLikeToRoute = async (id, like) =>
+  await Route.findByIdAndUpdate(id, like)
+
+export { postLikeToRoute }
